requirements.txt
fvcore==0.1.5.post20211023
iopath==0.1.9
numpy==1.21.3
opencv-python==4.5.4.58
packaging==21.0
Pillow==8.4.0
portalocker==2.3.2
pyparsing==3.0.3
PyYAML==6.0
tabulate==0.8.9
termcolor==1.1.0
torch==1.10.0
torchmetrics==0.6.0
torchvision==0.10.0
tqdm==4.62.3
typing-extensions==3.10.0.2
yacs==0.1.8
